requirements.txt
accelerate==1.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
cmake==3.30.4
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
diskcache==5.6.3
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
huggingface-hub==0.26.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
llama-cli==1.0.18
llama_cpp_python==0.3.1
MarkupSafe==3.0.1
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scipy==1.13.1
sentencepiece==0.2.0
six==1.16.0
sympy==1.13.1
tokenizers==0.20.1
torch==2.5.0+cu118
torchaudio==2.5.0+cu118
torchvision==0.20.0+cu118
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zipp==3.20.2
